requirements.txt
#requirements.txt
langchain==0.2.10
langchain-community==0.2.9
langchain-core==0.2.22
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
openai==1.59.8
python-dotenv==1.0.1
langchain_chroma==0.1.2
httpx==0.24.1

#.env

#OPENAI_API_KEY = ""
#LANGCHAIN_API_KEY=""
#LANGCHAIN_TRACING_V2=true
#LANGCHAIN_PROJECT=VECTORSTOREPROJECT

#.gitignore

#.env
#.venv


